move Ethlargement to NHMCore
