+++ NiceHashMiner.sln
@@ -48,8 +48,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "nvidiasetp0state", "src\Too
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FirewallRules", "src\Tools\FirewallRules\FirewallRules.csproj", "{BA2F08D2-DABD-4C2E-82B1-E1E5A3725911}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ethlargement", "src\Miners\Ethlargement\Ethlargement.csproj", "{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateLogReport", "src\Tools\CreateLogReport\CreateLogReport.csproj", "{663285C0-8F68-4638-BEB2-4BB042868445}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClaymoreDual14", "src\Miners\ClaymoreDual14\ClaymoreDual14.csproj", "{F28ECF66-1FD8-4A75-9E2B-D49F7036DE72}"
@@ -244,14 +242,6 @@ Global
 		{BA2F08D2-DABD-4C2E-82B1-E1E5A3725911}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BA2F08D2-DABD-4C2E-82B1-E1E5A3725911}.Release|x64.ActiveCfg = Release|Any CPU
 		{BA2F08D2-DABD-4C2E-82B1-E1E5A3725911}.Release|x64.Build.0 = Release|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Debug|x64.Build.0 = Debug|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Release|x64.ActiveCfg = Release|Any CPU
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A}.Release|x64.Build.0 = Release|Any CPU
 		{663285C0-8F68-4638-BEB2-4BB042868445}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{663285C0-8F68-4638-BEB2-4BB042868445}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{663285C0-8F68-4638-BEB2-4BB042868445}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -506,7 +496,6 @@ Global
 		{A95EE45C-D188-46E8-91F6-2CF858420BB5} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 		{9E106E41-0813-489D-A1B5-85ED434F2C3E} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
 		{BA2F08D2-DABD-4C2E-82B1-E1E5A3725911} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
-		{5F994AD1-02A2-4A00-90CA-5CD5943FD24A} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 		{663285C0-8F68-4638-BEB2-4BB042868445} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
 		{F28ECF66-1FD8-4A75-9E2B-D49F7036DE72} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 		{1B657592-641E-42EC-BC44-ABCAE62EF1FD} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
